refactor: merge interfaces and types folder
