+++ src/notifications.tsx
@@ -1,6 +1,6 @@
 import { List, Icon } from "@raycast/api";
 import { NotificationDropdown, NotificationMenu } from "./components/notifications";
-import { NotificationSortTypes } from "./types/notification-search";
+import { NotificationSortTypes } from "./types/sorts/notification-search";
 import { Fragment, useState } from "react";
 import { useNotifications } from "./hooks";
